requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit
flask

-e .